lib: Detach armv9a related code
