+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
+    "lib/armv9a",
     "lib/cose",
     "lib/rsi-el0",
     "lib/uart",
