+++ rmm/src/asm.rs
@@ -38,3 +38,17 @@ pub fn smc(cmd: usize, args: &[usize]) -> [usize; 8] {
 
     ret
 }
+
+#[inline(always)]
+pub fn dcache_flush(addr: usize, len: usize) {
+    let mut cur_addr = addr;
+    let addr_end = addr + len;
+    unsafe {
+        while cur_addr < addr_end {
+            asm!("dc civac, {}", in(reg) cur_addr);
+            asm!("dsb ish");
+            asm!("isb");
+            cur_addr += 64; // the cache line size is 64
+        }
+    }
+}
