+++ rmm/src/rmm/mod.rs
@@ -5,8 +5,8 @@ pub mod page;
 pub mod page_table;
 pub mod translation;
 
+use crate::asm::dcache_flush;
 use crate::config::PAGE_SIZE;
-use crate::helper::asm::dcache_flush;
 
 pub type PageMap = &'static dyn RmmPage;
 
@@ -41,4 +41,30 @@ impl RmmPage for MemoryMap {
         translation::unset_page_for_rmi(addr);
         true
     }
+}
+
+/// Call `rmm_exit` within `exception/vectors.s` and jumps to EL1.
+///
+/// Currently, this function gets [0usize; 3] as an argument to initialize
+/// x0, x1 and x2 registers.
+///
+/// When an exception occurs and the flow comes back to EL2 through `rmm_enter`,
+/// x0, x1 and x2 registers might be changed to contain additional information
+/// set from `handle_lower_exception`.
+/// These are the return values of this function.
+/// The return value encodes: [rmi::RET_XXX, ret_val1, ret_val2]
+/// In most cases, the function returns [rmi::RET_SUCCESS, _, _]
+/// pagefault returns [rmi::RET_PAGE_FAULT, faulted address, _]
+pub unsafe fn rmm_exit(args: [usize; 4]) -> [usize; 4] {
+    let mut ret: [usize; 4] = [0usize; 4];
+
+    core::arch::asm!(
+        "bl rmm_exit",
+        inlateout("x0") args[0] => ret[0],
+        inlateout("x1") args[1] => ret[1],
+        inlateout("x2") args[2] => ret[2],
+        inlateout("x3") args[3] => ret[3],
+    );
+
+    ret
 }
